update Flexophore similarity. Single node Flexophores are now allowed

+++ src/main/java/com/actelion/research/chem/descriptor/flexophore/completegraphmatcher/ObjectiveBlurFlexophoreHardMatchUncovered.java
@@ -854,7 +854,7 @@ public class ObjectiveBlurFlexophoreHardMatchUncovered implements IObjectiveComp
 		return score;
 	}
 
-	public float getSimilarityNodes(int indexNodeQuery, int indexNodeBase) {
+	public double getSimilarityNodes(int indexNodeQuery, int indexNodeBase) {
 		
 		if(arrSimilarityNodes[indexNodeQuery][indexNodeBase] < 0 || verbose){
